+++ CHANGES.md
@@ -46,6 +46,7 @@ UI:
 -   Show Database ownership information on dataset page (Admin Only)
 -   Added vocabulary suggestion for keywords & themes input on new dataset page
 -   Extracted keyword will be filtered by vocabulary APIs
+-   New add dataset page design
 
 Gateway: